New experimental right and middle click simulation (LP: #1191098).
Use button re-mapping like in Onboard 0.93 to allow for dragging.
This time the button mapping is restored with help from the XInput
event source. The 'GTK' event source still relies on click polling
doing the same. In either case, exiting or killing (SIGTERM) Onboard
will restore the buttons as well.