Support XDG Base Directory Specification (LP: #1074448).
Onboard's data directory moves from ~/.onboard to the XDG data directory,
which will usually be ~/.local/share/onboard.
Existing data in either ~/.onboard or ~/.sok is migrated to the new location.

System defaults are looked up in XDG_CONFIG_HOME, XDG_CONFIG_DIRS and the
previous locations, in this order.