Attempt to raise the keyboard over popups when force-to-top is enabled.
This was meant for bug LP #1404981, and it does indeed raise Onboard over the 
firefox URl-bar popup, but problems remain. The popup keeps getting events
that are meant for Onboard. However, it appears to work rather well in other 
applications, ZIM new page completion popup for example, so I'll leave it
enabled for now.