Allow left-clicking AppIndicator in KDE Plasma. Remove the indicator backend "StatusNotifier" again.
I had too many difficulties with implementing the dbusmenu for the
KDE StatusNotifier. Python3-dbus doesn't allow returning the recursive·
result of com.canonical.dbusmenu.GetLayout(). And using libdbusmenu
failed because the connection can't be shared with python3-dbus.

So there is now a dbus match for the Activate method when AppIndicator is
active. A hack that might fail when D-Bus security is enabled, 
but for now, in KDE Plasma, it seems to work.